Add MQTT/Home Assistant bridge on a persistent spa connection
Fixes:
- Power.Amps was missing the /10 scaling every other raw current/temperature
  field in this protocol uses, so it reported literal deciamps as amps (e.g.
  142 instead of 14.2A). Cross-checked against an independent power meter on
  the spa's own circuit to confirm the scale - the bogus value would read as
  142A on a circuit rated for 15A.

Persistent connection:
- Replace safespa's per-request dial/close hack with a single connection
  held for the process lifetime, guarded by the existing mutex.
- pkg/spanet gains NewWithDialer, which transparently redials and retries a
  command once if the connection is found dead, and SafeSpa.Do, so any
  caller - HTTP handlers, the MQTT bridge - queues onto the same connection
  instead of each dialing its own.
- Graceful shutdown: the first ctrl+c/SIGTERM drains in-flight HTTP
  requests and closes the spa connection cleanly; a second one force-exits
  immediately. Handled with our own signal channel rather than
  signal.NotifyContext's stop(), which doesn't reliably restore default
  signal behavior in practice.

MQTT:
- New pkg/hamqtt: a small, spa-agnostic Home Assistant MQTT discovery
  client - topics, discovery payloads, availability/LWT, command dispatch.
- New subcmd/server/mqttbridge: discovery entities for every setting and
  sensor the HTTP API exposes, all commands routed through SafeSpa.Do so
  MQTT and HTTP share the one spa connection safely.
- server gains -mqtt-broker/-mqtt-username/-mqtt-password/-mqtt-node-id/
  -mqtt-poll-interval flags to enable it.

Home Assistant specifics:
- water_heater entity for temperature control (current + target); its
  multi-topic shape doesn't fit the single state/command topic model every
  other entity uses, so pkg/hamqtt also gained lower-level Topic/PublishRaw/
  SubscribeRaw/PublishDiscovery primitives for it.
- Device availability now tracks actual spa reachability (goes offline
  after 3 consecutive failed polls), not just the MQTT broker connection.
- Apparent-power sensor (Volts x Amps, in VA) and a last-successful-update
  timestamp sensor.
- addon/spanet: add-on packaging (config.yaml, Dockerfile, run.sh
  translating options.json + Supervisor-injected MQTT service vars into
  flags), plus repository.yaml so this repo installs directly as a Home
  Assistant add-on repository.

+++ pkg/spanet/status.go
@@ -96,7 +96,7 @@ type TimeDate struct {
 
 type Power struct {
 	Volts        int     `rf:"R2,3"`
-	Amps         int     `rf:"R2,2"`
+	Amps         float64 `rf:"R2,2"` // *10
 	HeatingAmps  float64 `rf:"R3,23"`
 	CurrentLimit int     `rf:"R3,2"`
 	LoadShed     int     `rf:"R3,18"`
